+++ src/components/input/components/InputContainer.tsx
@@ -1,0 +1,5 @@
+import { ContainerProps } from '@/types';
+
+export const InputContainer = ({ children }: ContainerProps) => {
+  return <div className='input-container'>{children}</div>;
+};
