+++ src/components/input/components/InputTitle.tsx
@@ -1,0 +1,5 @@
+import { ContainerProps } from '@/types';
+
+export const InputTitle = ({ children }: ContainerProps) => {
+  return <span className='input-title'>{children}</span>;
+};
