+++ src/components/input/components/index.ts
@@ -1,0 +1,4 @@
+export * from './DefaultInput';
+export * from './InputContainer';
+export * from './InputBox';
+export * from './InputTitle';
