+++ src/types/component.ts
@@ -1,3 +1,5 @@
 export interface ContainerProps {
   children: React.ReactNode;
 }
+
+export type ClassNames = string[] | readonly string[];
