+++ src/utils/parser.ts
@@ -1,0 +1,7 @@
+import { ClassNames } from '@/types';
+
+export const Parser = {
+  classNames(classNames: ClassNames) {
+    return classNames.join(' ');
+  },
+};
